Fix issue reported at https://groups.google.com/forum/#!topic/webrtc-everywhere/Od8ezlwpPBI Adds support for x64 (Windows) Change Windows setup:  - to register the ActiveX under HKCU  - to install binaries under AppData folder instead of ProgramFiles  - to bundle 32bit and 64bit binaries

+++ common/_Common.cc
@@ -198,7 +198,7 @@ bool _File::Write(cpp11::shared_ptr<_Buffer>& buffer, bool append /*= false*/)
 				return false;
 			}
 		}
-		if (WriteFile(m_file, buffer->getPtr(), buffer->getSize(), &dwBytesWritten, &offset)) {
+		if (WriteFile(m_file, buffer->getPtr(), (DWORD)buffer->getSize(), &dwBytesWritten, &offset)) {
 			if (dwBytesWritten == buffer->getSize()) {
 				return (SetEndOfFile(m_file) == TRUE);
 			}
